# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agro API",
        "description": "Documentación de la API de autenticación con Swagger en Laravel 11",
        "contact": {
            "email": "soporte@tuempresa.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/about-us-home/{idAboutUsHome}": {
            "put": {
                "tags": [
                    "AboutUsHome"
                ],
                "summary": "Actualizar los datos de About Us Home",
                "operationId": "8381da444bf84f352799587cd1e7b1c4",
                "parameters": [
                    {
                        "name": "idAboutUsHome",
                        "in": "path",
                        "description": "ID de About Us Home",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "text_section_one",
                                    "text_section_two"
                                ],
                                "properties": {
                                    "text_section_one": {
                                        "type": "string",
                                        "example": "Nuestra misión es..."
                                    },
                                    "text_section_two": {
                                        "type": "string",
                                        "example": "Nuestro compromiso es..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Información actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Información actualizada correctamente"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "No se enviaron datos válidos para actualizar"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/about-us-home/image": {
            "post": {
                "tags": [
                    "AboutUsHome"
                ],
                "summary": "Actualizar la imagen de About Us Home",
                "operationId": "31af6878edebf1e1e41ce0e794ebea46",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "string",
                                        "format": "base64",
                                        "example": "data:image/png;base64,iVBORw0KGgoAAAANS..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Imagen actualizada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Error al actualizar la imagen"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}